[Favorite][ViewModel] Add ViewModel with mock data and search

+++ app/src/main/java/com/bif/locator/domain/model/Favorite.java
@@ -13,6 +13,6 @@ public class Favorite {
     public String address;
     public String description;
     public String notes;
-    public double rating;
+    public int rating;
     public String imagePath;
 }
